+++ docs/changelog.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * New vision example: MAML++. (@[DubiousCactus](https://github.com/DubiousCactus))
 * Add tutorial: "Demystifying Task Transforms", ([Varad Pimpalkhute](https://github.com/nightlessbaron/))
+* Add `l2l.nn.MetaModule` and `l2l.nn.ParameterTransform` for parameter-efficient finetuning.
+* Add `l2l.nn.freeze`and `l2l.nn.unfreeze`.
+* Add Adapters and LoRA examples.
 
 ### Changed